Add vector scalar mul
